feat: integrate circuit breaker into communicaManage::reinit()
- Add CircuitBreaker to communicate struct (per-device breaker)
- Check breaker.allowRequest() before reinit attempt
- Record failure when reinit returns fd < 0
- Record success and reset breaker on successful reinit
- Prevents retry storms when device is persistently unavailable

Benefits:
- Avoids infinite retry loops on broken devices
- Saves CPU and network resources
- Automatic recovery after cooldown period

+++ infra/circuit_breaker.h
@@ -65,7 +65,7 @@ public:
 
         if (state_ == CircuitState::Open) {
             // 检查冷却时间是否已过
-            if (getTickCount() - last_failure_time_ > cooldown_ms_) {
+            if (getTickCount() - last_failure_time_ > static_cast<uint64_t>(cooldown_ms_)) {
                 state_ = CircuitState::HalfOpen;
                 return true;
             }
